requirements.txt
python-dotenv
langchain-groq==0.3.2
langchain-postgres==0.0.14
langchain==0.3.23
langchain-community==0.3.21
langchain-huggingface==0.1.2
tiktoken==0.9.0
pypdf==5.4.0
pandas==2.2.3
openpyxl==3.1.5
fastapi==0.115.12
uvicorn==0.34.1
